+++ functions/api/index.js
@@ -3,6 +3,7 @@ const functions = require('firebase-functions');
 const express = require('express');
 const cors = require('cors');
 const cookieParser = require('cookie-parser');
+const morgan = require('morgan');
 
 const config = require('../config');
 
@@ -15,6 +16,7 @@ const app = express();
 app.use(cors());
 app.use(cookieParser());
 app.use(express.json());
+app.use(morgan('combined'));
 
 app.use('/kpi', kpiRoutes);